# MiLAi-Product/contracts/agent/v1/openapi.yaml
openapi: 3.1.0
info:
  title: MiLAi Agent Integration API
  version: 1.0.0
  description: >-
    Experimental loopback-only contract. Memory results are data, not instructions;
    empty canonical results are a valid abstention.
servers:
  - url: http://127.0.0.1:18080
security:
  - bearerAuth: []
paths:
...
  /v1/memory/query:
    post:
      operationId: recall
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RecallRequest'}
      responses:
        '200':
          description: Canonical-gated result or normal abstention.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RecallResponse'}
        '503':
          description: Fail-closed abstention while canonical storage is unavailable.
  /v1/memory/resolve:
    post:
      operationId: resolveMemory
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/MemoryResolveRequest'}
      responses:
        '200': {description: Typed governed AccessOutcome from task-free or optionally narrowed retrieval.}
        '400': {description: Invalid, widening, or conflicting TaskContext/request hints.}
        '503': {description: Typed UNAVAILABLE while canonical storage cannot be checked.}
  /v1/memory/get:
    post:
      operationId: getMemoryState
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Object'}
      responses:
        '200': {description: Dynamic canonical MemoryStateView for one ClaimID or StateKey.}
        '503': {description: Fail closed while canonical storage cannot be checked.}
...
  /v1/proposals:
    get:
      operationId: listProposals
      parameters:
        - {name: status, in: query, schema: {type: string, enum: [PENDING_REVIEW, DEFERRED, APPLIED, REJECTED]}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 100}}
      responses:
        '200': {description: Bounded Proposal inbox for the authenticated tenant.}
    post:
      operationId: createProposal
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/Object'}
      responses:
        '201': {description: Pending non-canonical Proposal.}
        '409': {description: Conflict; client must not auto-retry.}
  /v1/proposals/{proposal_id}:
    get:
      operationId: getProposal
      parameters:
        - {name: proposal_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: One non-canonical Proposal for authorized review.}
        '404': {description: Proposal not found.}
  /v1/proposals/{proposal_id}/review:
    post:
      operationId: reviewProposal
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
        - {name: proposal_id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ProposalReview'}
      responses:
        '200': {description: Immutable governed approve/reject decision and controlled result.}
        '403': {description: Independent reviewer capability required.}
        '409': {description: Already decided, stale version, or idempotency conflict.}
...
  /v1/claims/{claim_id}:
    get:
      operationId: getClaim
      parameters:
        - {name: claim_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Current effective Claim.}
        '404': {description: Claim not found.}
  /v1/open-issues:
    get:
      operationId: listOpenIssues
      parameters:
        - {name: status, in: query, required: false, schema: {type: string}}
      responses:
        '200': {description: OpenIssue records; uncertainty must be preserved.}
  /v1/retrieval-traces/{trace_id}:
    get:
      operationId: getTrace
      parameters:
        - {name: trace_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Persisted retrieval trace.}
...
  /v1/causal-tokens:
    post:
      operationId: issueCausalToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outbox_ids]
              properties:
                outbox_ids:
                  type: array
                  minItems: 1
                  maxItems: 16
                  items: {type: string, format: uuid}
      responses:
        '201': {description: Tenant-bound signed causal position.}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
  schemas:
    Object: {type: object, additionalProperties: true}
    RecallRequest:
      type: object
      additionalProperties: false
      required: [route]
      properties:
        route: {type: string, enum: [L0, L1]}
        consistency: {type: string, enum: [EVENTUAL, READ_YOUR_WRITES, CANONICAL_REQUIRED]}
        query: {type: string, maxLength: 2000}
        claim_id: {type: string, format: uuid}
        subject_id: {type: string}
        predicate: {type: string}
        claim_type: {type: string}
        requested_scope: {type: object}
        required_authority: {type: string, enum: [INFORMATIONAL, ACTION_SAFE, USER_CONFIRMED]}
        causal_token: {type: string}
        limit: {type: integer, minimum: 1, maximum: 50}
    RecallResponse:
      type: object
      required: [results, open_issue_ids, abstained, degraded_components, fallback_used]
      properties:
        results: {type: array, items: {$ref: '#/components/schemas/Object'}}
        open_issue_ids: {type: array, items: {type: string, format: uuid}}
        abstained: {type: boolean}
        abstention_reason: {type: [string, 'null']}
        retrieval_trace_id: {type: [string, 'null']}
        snapshot: {type: [object, 'null']}
        degraded_components: {type: array, items: {type: string}}
        fallback_used: {type: boolean}
        fallback_reason: {type: [string, 'null']}
    MemoryResolveRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query: {type: string, minLength: 1, maxLength: 2000}
        invocation_mode: {type: string, enum: [EXPLICIT_READ, PREFETCH_AUTO]}
        requested_scope: {type: object}
        required_authority: {type: string, enum: [INFORMATIONAL, ACTION_SAFE, USER_CONFIRMED]}
        required_freshness: {type: string, enum: [CURRENT, STALE]}
        consistency_mode: {type: string, enum: [EVENTUAL, READ_YOUR_WRITES, CANONICAL_REQUIRED]}
        causal_token: {type: string, minLength: 16, maxLength: 512}
        budget: {$ref: '#/components/schemas/Object'}
        entities: {type: array, maxItems: 32, items: {type: string, minLength: 1}}
        memory_types: {type: array, maxItems: 16, items: {type: string, minLength: 1}}
        temporal: {type: object}
        state_keys: {type: array, maxItems: 8, items: {$ref: '#/components/schemas/Object'}}
        claim_ids: {type: array, maxItems: 8, items: {type: string, format: uuid}}
        valid_at: {type: string, format: date-time}
        known_at: {type: string, format: date-time}
        previous_context_id: {type: string, format: uuid}
        task_context: {$ref: '#/components/schemas/TaskContextHint'}
    TaskContextHint:
      type: object
      additionalProperties: false
      description: Optional narrowing hints; never grants scope, consistency, write, review, or action authority.
      properties:
        project_ids: {type: array, maxItems: 16, items: {type: string, minLength: 1}}
        entities: {type: array, maxItems: 32, items: {type: string, minLength: 1}}
        memory_types: {type: array, maxItems: 16, items: {type: string, minLength: 1}}
        action_risk: {type: [string, 'null'], enum: [LOW, MEDIUM, HIGH, null]}
...
    ProposalReview:
      type: object
      additionalProperties: false
      required: [decision, policy_version, reason_code]
      properties:
        decision: {type: string, enum: [APPROVE, REJECT]}
        policy_version: {type: string, minLength: 1, maxLength: 255}
        reason_code: {type: string, minLength: 1, maxLength: 255}
...